+++ projects/app/src/pages/dataset/detail/components/Import/commonProgress/DataProcess.tsx
@@ -20,23 +20,20 @@ import LeftRadio from '@fastgpt/web/components/common/Radio/LeftRadio';
 import { TrainingTypeMap } from '@fastgpt/global/core/dataset/constants';
 import { ImportProcessWayEnum } from '@/web/core/dataset/constants';
 import MyTooltip from '@/components/MyTooltip';
-import { useImportStore } from '../Provider';
 import { useSystemStore } from '@/web/common/system/useSystemStore';
 import MyModal from '@fastgpt/web/components/common/MyModal';
 import { Prompt_AgentQA } from '@fastgpt/global/core/ai/prompt/agent';
 import Preview from '../components/Preview';
 import Tag from '@fastgpt/web/components/common/Tag/index';
+import { useContextSelector } from 'use-context-selector';
+import { DatasetImportContext } from '../Context';
 
-function DataProcess({
-  showPreviewChunks = true,
-  goToNext
-}: {
-  showPreviewChunks: boolean;
-  goToNext: () => void;
-}) {
+function DataProcess({ showPreviewChunks = true }: { showPreviewChunks: boolean }) {
   const { t } = useTranslation();
   const { feConfigs } = useSystemStore();
+
   const {
+    goToNext,
     processParamsForm,
     chunkSizeField,
     minChunkSize,
@@ -44,7 +41,7 @@ function DataProcess({
     showPromptInput,
     maxChunkSize,
     priceTip
-  } = useImportStore();
+  } = useContextSelector(DatasetImportContext, (v) => v);
   const { getValues, setValue, register } = processParamsForm;
   const [refresh, setRefresh] = useState(false);
