+++ projects/app/src/pages/dataset/detail/components/Import/commonProgress/PreviewData.tsx
@@ -2,15 +2,12 @@ import React from 'react';
 import Preview from '../components/Preview';
 import { Box, Button, Flex } from '@chakra-ui/react';
 import { useTranslation } from 'next-i18next';
+import { useContextSelector } from 'use-context-selector';
+import { DatasetImportContext } from '../Context';
 
-const PreviewData = ({
-  showPreviewChunks,
-  goToNext
-}: {
-  showPreviewChunks: boolean;
-  goToNext: () => void;
-}) => {
+const PreviewData = ({ showPreviewChunks }: { showPreviewChunks: boolean }) => {
   const { t } = useTranslation();
+  const goToNext = useContextSelector(DatasetImportContext, (v) => v.goToNext);
 
   return (
     <Flex flexDirection={'column'} h={'100%'}>
