+++ projects/app/src/pages/dataset/detail/components/Import/components/PreviewChunks.tsx
@@ -4,11 +4,12 @@ import { ImportSourceItemType } from '@/web/core/dataset/type';
 import { useQuery } from '@tanstack/react-query';
 import MyRightDrawer from '@fastgpt/web/components/common/MyDrawer/MyRightDrawer';
 import { getPreviewChunks } from '@/web/core/dataset/api';
-import { useImportStore } from '../Provider';
 import { ImportDataSourceEnum } from '@fastgpt/global/core/dataset/constants';
 import { splitText2Chunks } from '@fastgpt/global/common/string/textSplitter';
 import { useToast } from '@fastgpt/web/hooks/useToast';
 import { getErrText } from '@fastgpt/global/common/error/utils';
+import { useContextSelector } from 'use-context-selector';
+import { DatasetImportContext } from '../Context';
 
 const PreviewChunks = ({
   previewSource,
@@ -18,7 +19,10 @@ const PreviewChunks = ({
   onClose: () => void;
 }) => {
   const { toast } = useToast();
-  const { importSource, chunkSize, chunkOverlapRatio, processParamsForm } = useImportStore();
+  const { importSource, chunkSize, chunkOverlapRatio, processParamsForm } = useContextSelector(
+    DatasetImportContext,
+    (v) => v
+  );
 
   const { data = [], isLoading } = useQuery(
     ['previewSource'],
