+++ projects/app/src/pages/dataset/detail/components/Import/components/PreviewRawText.tsx
@@ -4,10 +4,11 @@ import { ImportSourceItemType } from '@/web/core/dataset/type';
 import { useQuery } from '@tanstack/react-query';
 import { getPreviewFileContent } from '@/web/common/file/api';
 import MyRightDrawer from '@fastgpt/web/components/common/MyDrawer/MyRightDrawer';
-import { useImportStore } from '../Provider';
 import { ImportDataSourceEnum } from '@fastgpt/global/core/dataset/constants';
 import { useToast } from '@fastgpt/web/hooks/useToast';
 import { getErrText } from '@fastgpt/global/common/error/utils';
+import { useContextSelector } from 'use-context-selector';
+import { DatasetImportContext } from '../Context';
 
 const PreviewRawText = ({
   previewSource,
@@ -17,7 +18,7 @@ const PreviewRawText = ({
   onClose: () => void;
 }) => {
   const { toast } = useToast();
-  const { importSource } = useImportStore();
+  const { importSource } = useContextSelector(DatasetImportContext, (v) => v);
 
   const { data, isLoading } = useQuery(
     ['previewSource', previewSource?.dbFileId],
